+++ example/lib/main.dart
@@ -1,7 +1,7 @@
+import 'dart:async';
 import 'package:flutter/material.dart';
-import 'package:jh_debug/jh_debug.dart';
+import 'package:jh_debug/jh_debug.dart' show DebugMode, jhDebug, jhDebugMain;
 import 'TestPage.dart';
-
 import 'DetailsPage.dart';
 
 void main() {
@@ -9,6 +9,20 @@ void main() {
     appChild: MyApp(),
     debugMode: DebugMode.inConsole,
     errorCallback: (FlutterErrorDetails) {},
+    beforeAppChildFn: (() async {
+      Completer<void> completer = Completer<void>(); // 创建一个Completer对象
+
+      // 模拟定时器处理，假设需要不定时间处理
+      Timer(const Duration(seconds: 3), () {
+        print('定时器处理完成');
+        completer.complete(); // 完成Completer，通知任务完成
+      });
+
+      // 在Completer完成前等待
+      await completer.future;
+
+      print('执行前!');
+    }),
   );
 }
 
@@ -29,7 +43,7 @@ class _MyAppState extends State<MyApp> {
     return MaterialApp(
       // home: TestPage(),
       // showPerformanceOverlay: true,
-      theme: ThemeData.dark(),
+      theme: ThemeData.light(),
       navigatorKey: jhDebug.getNavigatorKey,
       routes: {
         '/': (BuildContext context) => TestPage(),