+++ example/lib/services/api.dart
@@ -1,7 +1,8 @@
+import '../models/common.m.dart';
 import '../utils/request.dart';
 
 /// 获取APP最新版本号, 演示更新APP组件
-Future<Map> getNewVersion() async {
+Future<NewVersionData> getNewVersion() async {
   var resData = await Request.get(
     'https://www.fastmock.site/mock/4fa6906a0c87f7c2513e85cb98eb4bdb/shop/nCoV/api/area?key=778',
     queryParameters: {
@@ -11,5 +12,5 @@ Future<Map> getNewVersion() async {
   );
   print('请示后》》》');
 
-  return resData ?? {};
+  return (resData.data ?? {}) as NewVersionData;
 }