+++ example/lib/utils/dio/dio_error_util.dart
@@ -1,29 +1,33 @@
 import 'package:dio/dio.dart';
 
 class DioErrorUtil {
-  static String handleError(DioError? error) {
+  static String handleError(DioException? error) {
     String errorDescription = "";
-    if (error is DioError) {
+    if (error is DioException) {
       switch (error.type) {
-        case DioErrorType.cancel:
+        case DioExceptionType.cancel:
           errorDescription = "Request to API server was cancelled";
           break;
-        case DioErrorType.connectTimeout:
+        case DioExceptionType.connectionTimeout:
           errorDescription = "Connection timeout with API server";
           break;
-        case DioErrorType.other:
+        case DioExceptionType.unknown:
           errorDescription =
               "Connection to API server failed due to internet connection";
           break;
-        case DioErrorType.receiveTimeout:
+        case DioExceptionType.receiveTimeout:
           errorDescription = "Receive timeout in connection with API server";
           break;
-        case DioErrorType.response:
+        case DioExceptionType.badResponse:
           errorDescription =
               "Received invalid status code: ${error.response?.statusCode}";
           break;
-        case DioErrorType.sendTimeout:
+        case DioExceptionType.sendTimeout:
           errorDescription = "Send timeout in connection with API server";
+          break;
+        case DioExceptionType.badCertificate:
+          break;
+        case DioExceptionType.connectionError:
           break;
       }
     } else {